Modifications to CholmodSupport.h for lme4
In earlier versions of RcppEigen I made some modifications in the
CholmodSupport.h file for specific tasks in lme4 (SolveInPlace) and
for the `wrap` capabilities in RcppEigen itself.  These are now installed
in this version but the objective is to remove them when feasible.

Also updated the version and date in the DESCRIPTION file and changed
the dependence to R (>= 3.0.0)

The reason setting R (>= 3.0.2) in the previous release was to avoid a
NOTE in R CMD check.